+++ R/filter.R
@@ -7,18 +7,19 @@
 #'
 #' @return A \code{gho} object.
 #' @export
-filter_gho <- function (x, ...) {
-  .dots <- lazyeval::lazy_dots(...)
-  filter_gho_(x, .dots = .dots)
-}
+# filter_gho <- function (x, ...) {
+#   .dots <- lazyeval::lazy_dots(...)
+#   filter_gho_(x, .dots = ...)
+# }
 
-filter_gho_ <- function(x, .dots) {
+filter_gho <- function(x, ...) {
   if (is.null(gho_attr <- attr(x, "attrs"))) {
     stop("Attempt to filter a GHO object with no attribute.")
   }
 
+  .dots <- rlang::quos(...)
   filtered_attrs <- gho_attr %>%
-    dplyr::filter_(.dots = .dots)
+    dplyr::filter(!!!.dots)
 
   codes <- unique(filtered_attrs$code)